+++ source/poHelpers.cpp
@@ -228,7 +228,7 @@ namespace po {
     
     
     //------------------------------------------------------------------------
-    std::string base64_encode(unsigned char const* bytes_to_encode, unsigned int in_len) {
+    std::string base64_encode(unsigned char const* bytes_to_encode, size_t in_len) {
         std::string ret;
         int i = 0;
         int j = 0;
@@ -273,7 +273,7 @@ namespace po {
     
     //------------------------------------------------------------------------
     std::string base64_decode(std::string const& encoded_string) {
-        int in_len = encoded_string.size();
+        size_t in_len = encoded_string.size();
         int i = 0;
         int j = 0;
         int in_ = 0;
